specify type cast

+++ exercises/temperature.c
@@ -10,7 +10,7 @@ main() {
 	step = 20;
 	
 	fahr = (float) lower;
-	while (fahr <= upper) {
+	while (fahr <= (float) upper) {
 		celsius = 5.0 * (fahr-32.0) / 9.0;
 		printf("%3.0f %6.1f\n", fahr, celsius);
 		fahr = fahr + step;
